Move RoutablePoint to common module; Add routable points for OfflineSearchResult
