+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.0.0-beta.36-SNAPSHOT
 
 ### Breaking changes
-- [CORE] `AsyncOperationTask`, `SearchCancellationException`, `SearchRequestException` have been moved to `com.mapbox.search.common` package. `MainThreadWorker`, and `SearchSdkMainThreadWorker` have been moved to `com.mapbox.search.common.concurrent`.
+- [CORE] `AsyncOperationTask`, `SearchCancellationException`, `SearchRequestException`, `RoutablePoint` have been moved to `com.mapbox.search.common` package. `MainThreadWorker`, and `SearchSdkMainThreadWorker` have been moved to `com.mapbox.search.common.concurrent`.
 - [CORE] `SearchRequestTask` has been replaced with `AsyncOperationTask`
 - [CORE] `SearchSuggestion` is a class now. Functions signature of `SearchSuggestionsCallback`, `SearchSelectionCallback`, `SearchMultipleSelectionCallback` have been changed for Java users.
 - [CORE] Offline functionality has been moved to a separate dependency which can be resolved via `com.mapbox.search:offline:$version`. At the moment Offline SDK version is the same as the main Search SDK version. `OfflineSearchEngine` instance can be retrieved via `OfflineSearchEngine.create(OfflineSearchEngineSettings)`.
