+++ MapboxSearch/offline/api/api-metalava.txt
@@ -167,6 +167,7 @@ package com.mapbox.search.offline {
     method public Double? getDistanceMeters();
     method public String getId();
     method public String getName();
+    method public java.util.List<com.mapbox.search.common.RoutablePoint>? getRoutablePoints();
     method public com.mapbox.search.offline.OfflineSearchResultType getType();
     property public final com.mapbox.search.offline.OfflineSearchAddress? address;
     property public final com.mapbox.geojson.Point coordinate;
@@ -174,6 +175,7 @@ package com.mapbox.search.offline {
     property public final Double? distanceMeters;
     property public final String id;
     property public final String name;
+    property public final java.util.List<com.mapbox.search.common.RoutablePoint>? routablePoints;
     property public final com.mapbox.search.offline.OfflineSearchResultType type;
   }
