+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/favorite/FavoriteRecordsSerializerTest.kt
@@ -5,6 +5,7 @@ import com.mapbox.search.BaseTest
 import com.mapbox.search.ImageInfo
 import com.mapbox.search.SearchResultMetadata
 import com.mapbox.search.TestData
+import com.mapbox.search.common.RoutablePoint
 import com.mapbox.search.metadata.OpenHours
 import com.mapbox.search.metadata.OpenPeriod
 import com.mapbox.search.metadata.ParkingData
@@ -12,7 +13,6 @@ import com.mapbox.search.metadata.WeekDay
 import com.mapbox.search.metadata.WeekTimestamp
 import com.mapbox.search.record.FavoriteRecord
 import com.mapbox.search.record.FavoriteRecordsSerializer
-import com.mapbox.search.result.RoutablePoint
 import com.mapbox.search.result.SearchAddress
 import com.mapbox.search.result.SearchResultType
 import org.junit.Assert
