+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/record/FavoriteRecord.kt
@@ -3,7 +3,7 @@ package com.mapbox.search.record
 import android.os.Parcelable
 import com.mapbox.geojson.Point
 import com.mapbox.search.SearchResultMetadata
-import com.mapbox.search.result.RoutablePoint
+import com.mapbox.search.common.RoutablePoint
 import com.mapbox.search.result.SearchAddress
 import com.mapbox.search.result.SearchResultType
 import kotlinx.parcelize.Parcelize
