+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/record/HistoryDataProvider.kt
@@ -9,6 +9,7 @@ import com.mapbox.search.base.result.BaseSearchResult
 import com.mapbox.search.base.task.AsyncOperationTaskImpl
 import com.mapbox.search.base.utils.LocalTimeProvider
 import com.mapbox.search.base.utils.TimeProvider
+import com.mapbox.search.base.utils.extension.mapToPlatform
 import com.mapbox.search.common.AsyncOperationTask
 import com.mapbox.search.result.mapToPlatform
 import java.util.PriorityQueue
