+++ MapboxSearch/sdk/src/sharedTest/java/com/mapbox/search/tests_support/SdkTestDataFactory.kt
@@ -14,11 +14,12 @@ import com.mapbox.search.base.result.BaseSearchSuggestion
 import com.mapbox.search.base.result.BaseServerSearchSuggestion
 import com.mapbox.search.base.result.BaseSuggestAction
 import com.mapbox.search.base.result.SearchRequestContext
+import com.mapbox.search.base.utils.extension.mapToCore
+import com.mapbox.search.common.RoutablePoint
 import com.mapbox.search.common.createTestCoreRequestOptions
 import com.mapbox.search.record.FavoriteRecord
 import com.mapbox.search.record.HistoryRecord
 import com.mapbox.search.result.ResultAccuracy
-import com.mapbox.search.result.RoutablePoint
 import com.mapbox.search.result.SearchAddress
 import com.mapbox.search.result.SearchResult
 import com.mapbox.search.result.SearchResultType
