+++ MapboxSearch/ui/api/api-metalava.txt
@@ -139,7 +139,7 @@ package com.mapbox.search.ui.view.place {
   }
 
   @kotlinx.parcelize.Parcelize public final class SearchPlace implements android.os.Parcelable {
-    ctor public SearchPlace(String id, String name, String? descriptionText, com.mapbox.search.result.SearchAddress? address, java.util.List<? extends com.mapbox.search.result.SearchResultType> resultTypes, com.mapbox.search.record.IndexableRecord? record, com.mapbox.geojson.Point coordinate, java.util.List<com.mapbox.search.result.RoutablePoint>? routablePoints, java.util.List<java.lang.String>? categories, String? makiIcon, com.mapbox.search.SearchResultMetadata? metadata, Double? distanceMeters, com.mapbox.search.ui.view.place.IncorrectSearchPlaceFeedback? feedback);
+    ctor public SearchPlace(String id, String name, String? descriptionText, com.mapbox.search.result.SearchAddress? address, java.util.List<? extends com.mapbox.search.result.SearchResultType> resultTypes, com.mapbox.search.record.IndexableRecord? record, com.mapbox.geojson.Point coordinate, java.util.List<com.mapbox.search.common.RoutablePoint>? routablePoints, java.util.List<java.lang.String>? categories, String? makiIcon, com.mapbox.search.SearchResultMetadata? metadata, Double? distanceMeters, com.mapbox.search.ui.view.place.IncorrectSearchPlaceFeedback? feedback);
     method public static com.mapbox.search.ui.view.place.SearchPlace createFromIndexableRecord(com.mapbox.search.record.IndexableRecord record, com.mapbox.geojson.Point coordinate, Double? distanceMeters);
     method public static com.mapbox.search.ui.view.place.SearchPlace createFromSearchResult(com.mapbox.search.result.SearchResult searchResult, com.mapbox.search.ResponseInfo responseInfo, com.mapbox.geojson.Point coordinate, Double? distanceMeters = searchResult.distanceMeters);
     method public static com.mapbox.search.ui.view.place.SearchPlace createFromSearchResult(com.mapbox.search.result.SearchResult searchResult, com.mapbox.search.ResponseInfo responseInfo, com.mapbox.geojson.Point coordinate);
@@ -155,7 +155,7 @@ package com.mapbox.search.ui.view.place {
     method public String getName();
     method public com.mapbox.search.record.IndexableRecord? getRecord();
     method public java.util.List<com.mapbox.search.result.SearchResultType> getResultTypes();
-    method public java.util.List<com.mapbox.search.result.RoutablePoint>? getRoutablePoints();
+    method public java.util.List<com.mapbox.search.common.RoutablePoint>? getRoutablePoints();
     property public final com.mapbox.search.result.SearchAddress? address;
     property public final java.util.List<java.lang.String>? categories;
     property public final com.mapbox.geojson.Point coordinate;
@@ -168,7 +168,7 @@ package com.mapbox.search.ui.view.place {
     property public final String name;
     property public final com.mapbox.search.record.IndexableRecord? record;
     property public final java.util.List<com.mapbox.search.result.SearchResultType> resultTypes;
-    property public final java.util.List<com.mapbox.search.result.RoutablePoint>? routablePoints;
+    property public final java.util.List<com.mapbox.search.common.RoutablePoint>? routablePoints;
   }
 
   public final class SearchPlaceBottomSheetView extends com.mapbox.search.ui.view.SearchSdkFrameLayout implements androidx.coordinatorlayout.widget.CoordinatorLayout.AttachedBehavior {
